services/auth-service/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
redis>=5.0,<6.0
httpx>=0.28,<1.0
pytest>=8.0,<9.0
pytest-asyncio>=0.25,<1.0
bcrypt>=4.0,<6.0
python-jose[cryptography]>=3.3,<4.0
aiosqlite>=0.20,<1.0
fakeredis>=2.21,<3.0
